// src/models/CharacterModel.ts
export interface Character {
  _id?: string;
  owner?: {
    username: string;
    email: string;
  };
  name: string;
  race: string;
  class: string;
  level: number;
  healthPoints: number;
  background: string;
  stats: {
    strength: number;
    dexterity: number;
    constitution: number;
    intelligence: number;
    wisdom: number;
    charisma: number;
  };
  skills: {
    acrobatics: number;
    animalHandling: number;
    arcana: number;
    athletics: number;
    deception: number;
    history: number;
    insight: number;
    intimidation: number;
    investigation: number;
    medicine: number;
    nature: number;
    perception: number;
    performance: number;
    persuasion: number;
    religion: number;
    sleightOfHand: number;
    stealth: number;
    survival: number;
  };
  equipment: string[];
  spells: string[];
  features: string[];
  notes: string;
  createdAt?: string;
  updatedAt?: string;
  __v?: number;
}